Prep for 4.0.0 beta
Update NodeJS versions
